map basic check done, next to check map solvability

+++ libft/ft_strlen.c
@@ -14,8 +14,8 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
-
+	size_t	i;
+	
 	i = 0;
 	while (s[i] != '\0')
 		i++;
